feat: add selectAggregatedEarnUserLpOpportunities selector
