+++ src/state/slices/opportunitiesSlice/selectors/index.ts
@@ -1,0 +1,2 @@
+export * from './lpSelectors'
+export * from './stakingSelectors'
